#228 : don't classify entity added in the brief section of the NC and NPD processes

+++ becpg-core/src/main/java/fr/becpg/repo/workflow/activiti/nc/CreateNC.java
@@ -127,17 +127,14 @@ public class CreateNC extends BaseJavaDelegate {
 						nodeService.createAssociation(ncNodeRef, node.getNodeRef(), BeCPGModel.ASSOC_CLIENTS);
 					}
 
-					// Move file from pkgNodeRef
+					// Move documents from pkgNodeRef
 					List<FileInfo> files = fileFolderService.listFiles(pkgNodeRef);
 					NodeRef briefNodeRef = getDocumentsFolder(ncNodeRef);
 					for (FileInfo file : files) {
 						String name = (String) nodeService.getProperty(file.getNodeRef(), ContentModel.PROP_NAME);
-						if (briefNodeRef != null) {
+						if (briefNodeRef != null && nodeService.getType(briefNodeRef).equals(ContentModel.TYPE_CONTENT)) {
 							fileFolderService.move(file.getNodeRef(), briefNodeRef, name);
 							nodeService.removeChild(pkgNodeRef, file.getNodeRef());
-						} else {
-							logger.error("No documents folder found");
-							break;
 						}
 					}
